refactor: centralize role flag helpers
